+++ eagleai/templates/login.html
@@ -1,0 +1,72 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <!-- Bootstrap CSS-->
+    <link
+      rel="stylesheet"
+      href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css"
+      integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
+      crossorigin="anonymous"
+    />
+    <link
+      rel="stylesheet"
+      href="{{ url_for('static',    filename='css/mystyle.css') }}"
+    />
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+    <link
+      href="https://fonts.googleapis.com/css?family=Open Sans"
+      rel="stylesheet"
+    />
+    <title>Login</title>
+  </head>
+
+  <body style="background-color: #eaeaea">
+    <div class="container-fluid">
+      <div
+        class="top"
+        style="
+          background-color: #73131d;
+          height: 140px;
+          display: block;
+          unicode-bidi: isolate;
+        "
+      >
+        <div class="bg-image"></div>
+      </div>
+
+      <section class="login">
+        <h3 class="">Sign in to continue</h3>
+        <form class="row mx-auto g-3">
+          <div class="col-auto mx-auto my-1">
+            <input
+              type="text"
+              class="form-control"
+              id="inputUsername"
+              placeholder="Username"
+            />
+          </div>
+          <div class="col-auto mx-auto my-1">
+            <input
+              type="password"
+              class="form-control"
+              id="inputPassword"
+              placeholder="Password"
+            />
+          </div>
+          <div class="col-auto mx-auto my-auto">
+            <button
+              type="submit"
+              class="btn btn-primary mb-3 my-3"
+              style="background-color: #73131d"
+            >
+              Sign In
+            </button>
+            <a class="nav-link" href="{{ url_for('views.home') }}"
+              >Fake Login</a
+            >
+          </div>
+        </form>
+      </section>
+    </div>
+  </body>
+</html>
